refactor(ui): extract launchAssistantChat shared helper
